+++ CA7/DeviceTree/WildhornAV/kernel/stm32mp157d-wildhornav-mx.dts
@@ -58,7 +58,7 @@
 			no-map;
 		};
 		/* RETRAM memory region reserved for firmware code and data */
-			retram: retram@0x38000000 {
+		retram: retram@0x38000000 {
 			reg = <0x38000000 0x10000>;
 		};
 		/* MCUSRAM memory region reserved for firmware code and data */
@@ -232,15 +232,6 @@
 			pinmux = <STM32_PINMUX('B', 0, RSVD)>, /* TIM3_CH3 */
 					 <STM32_PINMUX('C', 6, RSVD)>, /* TIM3_CH1 */
 					 <STM32_PINMUX('C', 7, RSVD)>; /* TIM3_CH2 */
-		};
-	};
-
-	m4_tim4_pwm_pins_mx: m4_tim4_pwm_mx-0 {
-		pins {
-			pinmux = <STM32_PINMUX('B', 6, RSVD)>, /* TIM4_CH1 */
-					 <STM32_PINMUX('D', 13, RSVD)>, /* TIM4_CH2 */
-					 <STM32_PINMUX('D', 14, RSVD)>, /* TIM4_CH3 */
-					 <STM32_PINMUX('D', 15, RSVD)>; /* TIM4_CH4 */
 		};
 	};
 
@@ -676,8 +667,6 @@
 };
 
 &m4_timers4{
-	pinctrl-names = "default";
-	pinctrl-0 = <&m4_tim4_pwm_pins_mx>;
 	status = "okay";
 
 	/* USER CODE BEGIN m4_timers4 */
