+++ CA7/DeviceTree/WildhornAV/u-boot/stm32mp157d-wildhornav-mx.dts.bak
@@ -803,6 +803,20 @@
 	/* USER CODE END uart8 */
 };
 
+&usbh_ehci{
+	status = "okay";
+
+	/* USER CODE BEGIN usbh_ehci */
+	/* USER CODE END usbh_ehci */
+};
+
+&usbh_ohci{
+	status = "okay";
+
+	/* USER CODE BEGIN usbh_ohci */
+	/* USER CODE END usbh_ohci */
+};
+
 &usbotg_hs{
 	u-boot,dm-pre-reloc;
 	status = "okay";
