+++ CM4/RemoteProc/stop_fw.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+
+
+echo -n stop > /sys/class/remoteproc/remoteproc0/state
