+++ CM4/ert/abstraction/gpio.c
@@ -1,25 +1,20 @@
-/*  Title		: Control
- *  Filename	: control.c
+/*  Title		: GPIO
+ *  Filename	: gpio.c
  *	Author		: iacopo sprenger
- *	Date		: 20.01.2022
+ *	Date		: 30.03.2021
  *	Version		: 0.1
- *	Description	: control
+ *	Description	: GPIO hardware abstraction
  */
 
 /**********************
  *	INCLUDES
  **********************/
 
-#include <main.h>
-#include <cmsis_os.h>
-#include <control.h>
-#include <led.h>
+#include "gpio.h"
 
 /**********************
  *	CONSTANTS
  **********************/
-
-#define CONTROL_HEART_BEAT	100
 
 
 /**********************
@@ -45,23 +40,5 @@
 /**********************
  *	DECLARATIONS
  **********************/
-
-void control_thread(void * arg) {
-	static TickType_t last_wake_time;
-	static const TickType_t period = pdMS_TO_TICKS(CONTROL_HEART_BEAT);
-
-	led_init();
-
-	led_set_color(LED_BLUE);
-
-	last_wake_time = xTaskGetTickCount();
-
-	for(;;) {
-
-
-		vTaskDelayUntil( &last_wake_time, period );
-	}
-}
-
 
 /* END */
