+++ CM4/ert/template/template.c
