+++ WildhornAV.ioc
@@ -558,8 +558,10 @@ PB15.Locked=true
 PB15.Signal=SPI2_MOSI
 PB5.Locked=true
 PB5.Signal=SPI1_MOSI
+PB6.GPIOParameters=GPIO_Label
+PB6.GPIO_Label=S1_GP1
 PB6.Locked=true
-PB6.Signal=S_TIM4_CH1
+PB6.Signal=GPIO_Output
 PB7.GPIOParameters=GPIO_Speed
 PB7.GPIO_Speed=GPIO_SPEED_FREQ_VERY_HIGH
 PB7.Locked=true
@@ -623,12 +625,18 @@ PD10.Signal=SPI3_MISO
 PD12.Locked=true
 PD12.Mode=I2C
 PD12.Signal=I2C1_SCL
+PD13.GPIOParameters=GPIO_Label
+PD13.GPIO_Label=S1_GP2
 PD13.Locked=true
-PD13.Signal=S_TIM4_CH2
+PD13.Signal=GPIO_Output
+PD14.GPIOParameters=GPIO_Label
+PD14.GPIO_Label=S1_GP3
 PD14.Locked=true
-PD14.Signal=S_TIM4_CH3
+PD14.Signal=GPIO_Output
+PD15.GPIOParameters=GPIO_Label
+PD15.GPIO_Label=S1_GP4
 PD15.Locked=true
-PD15.Signal=S_TIM4_CH4
+PD15.Signal=GPIO_Output
 PD2.GPIOParameters=GPIO_PuPd
 PD2.GPIO_PuPd=GPIO_PULLUP
 PD2.Locked=true
@@ -857,14 +865,6 @@ SH.S_TIM3_CH3.0=TIM3_CH3,PWM Generation3 CH3
 SH.S_TIM3_CH3.ConfNb=1
 SH.S_TIM3_CH4.0=TIM3_CH4
 SH.S_TIM3_CH4.ConfNb=1
-SH.S_TIM4_CH1.0=TIM4_CH1,PWM Generation1 CH1
-SH.S_TIM4_CH1.ConfNb=1
-SH.S_TIM4_CH2.0=TIM4_CH2,PWM Generation2 CH2
-SH.S_TIM4_CH2.ConfNb=1
-SH.S_TIM4_CH3.0=TIM4_CH3,PWM Generation3 CH3
-SH.S_TIM4_CH3.ConfNb=1
-SH.S_TIM4_CH4.0=TIM4_CH4,PWM Generation4 CH4
-SH.S_TIM4_CH4.ConfNb=1
 SH.S_TIM5_CH1.0=TIM5_CH1,PWM Generation1 CH1
 SH.S_TIM5_CH1.ConfNb=1
 SH.S_TIM5_CH2.0=TIM5_CH2,PWM Generation2 CH2
@@ -877,11 +877,6 @@ TIM3.Channel-PWM\ Generation1\ CH1=TIM_CHANNEL_1
 TIM3.Channel-PWM\ Generation2\ CH2=TIM_CHANNEL_2
 TIM3.Channel-PWM\ Generation3\ CH3=TIM_CHANNEL_3
 TIM3.IPParameters=Channel-PWM Generation1 CH1,Channel-PWM Generation2 CH2,Channel-PWM Generation3 CH3
-TIM4.Channel-PWM\ Generation1\ CH1=TIM_CHANNEL_1
-TIM4.Channel-PWM\ Generation2\ CH2=TIM_CHANNEL_2
-TIM4.Channel-PWM\ Generation3\ CH3=TIM_CHANNEL_3
-TIM4.Channel-PWM\ Generation4\ CH4=TIM_CHANNEL_4
-TIM4.IPParameters=Channel-PWM Generation1 CH1,Channel-PWM Generation2 CH2,Channel-PWM Generation3 CH3,Channel-PWM Generation4 CH4
 TIM5.Channel-PWM\ Generation1\ CH1=TIM_CHANNEL_1
 TIM5.Channel-PWM\ Generation2\ CH2=TIM_CHANNEL_2
 TIM5.Channel-PWM\ Generation3\ CH3=TIM_CHANNEL_3
